+++ gulpfile.ts
@@ -634,8 +634,15 @@ namespace Gulp {
   (gulp as any).task('gal-rel:process-readme', ['gal-rel:create-zip'],
     (cb) => {
       printOptions();
-      BuildGalRel.buildReadMe();
-      cb();
+      BuildGalRel.buildIndexHtml(BuildGalRel.buildReadMe());
+
+      return gulp.src([
+        `./node_modules/highlight.js/styles/github.css`,
+        `./node_modules/font-awesome/css/font-awesome.css`,
+        './node_modules/highlight.js/lib/highlight.js',
+      ])
+        .pipe(gulp.dest(BuildGalRel.DST_GALLERY_RELEASE_PATH + '/node_modules'))
+        .pipe(gulpPreserveTime());
     });